monitoring/requirements.txt
kafka-python==2.0.2
prometheus_client==0.15.0
